Triangle Minesweeper in almost done.
Made progress on rednering the board and placing everything properly inside the triangles. Also able to check if I'm clicking in a triangle. Last thing before completion is 3 funtions. Need to place bombs. Set counters +1 if bomb next to it. and lastly the recursive call that clicks multiple boxes if blank